[ci skip] refactor file upload
