+++ src/services/fileStorage/hooks/model-hooks.js
@@ -1,0 +1,25 @@
+'use strict';
+
+const globalHooks = require('../../../hooks');
+const hooks = require('feathers-hooks');
+const auth = require('feathers-authentication');
+
+exports.before = {
+	all: [auth.hooks.authenticate('jwt')],
+	find: [],
+	get: [],
+	create: [],
+	update: [],
+	patch: [],
+	remove: []
+};
+
+exports.after = {
+	all: [],
+	find: [],
+	get: [],
+	create: [],
+	update: [],
+	patch: [],
+	remove: []
+};
