+++ src/services/fileStorage/strategies/awsS3.js
@@ -7,6 +7,7 @@ const fs = require('fs');
 const pathUtil = require('path').posix;
 const logger = require('winston');
 const filePermissionHelper = require('../utils/filePermissionHelper');
+const removeLeadingSlash = require('../utils/leadingSlashHelper');
 let awsConfig;
 try {
 	awsConfig = require("../../../../config/secrets.json").aws;
@@ -41,10 +42,10 @@ const splitFilesAndDirectories = (path, data) => {
 		const relativePath = removeLeadingSlash(entry.key.replace(path, ''));
 		const pathComponents = relativePath.split('/');
 
-		if (pathComponents.length == 1) {
+		if (pathComponents.length === 1) {
 			files.push(entry);
 		} else {
-			if(entry.name == ".scfake") {	// prevent duplicates showing up by only considering .scfake
+			if(entry.name === ".scfake") {	// prevent duplicates showing up by only considering .scfake
 				const components = entry.key.split('/');
 				const directoryName = components[components.length - 2];	// the component before '.scfake'
 				directories.push({
@@ -61,11 +62,6 @@ const splitFilesAndDirectories = (path, data) => {
 		files,
 		directories
 	};
-};
-
-const removeLeadingSlash = path => {
-	if (path[0] == '/') path = path.substring(1);
-	return path;
 };
 
 const getFileMetadata = (storageContext, awsObjects, bucketName, s3) => {
@@ -178,37 +174,32 @@ class AWSS3Strategy extends AbstractFileStorageStrategy {
 	}
 
 	generateSignedUrl(userId, path, fileType, action) {
-		if (!userId || !path || !action || (action == 'putObject' && !fileType)) return Promise.reject(new errors.BadRequest('Missing parameters'));
-		path = removeLeadingSlash(pathUtil.normalize(path));	// remove leading and double slashes
+		if (!userId || !path || !action || (action === 'putObject' && !fileType)) return Promise.reject(new errors.BadRequest('Missing parameters'));
 		const fileName = pathUtil.basename(path);
 		let dirName = pathUtil.dirname(path);
+		return UserModel.findById(userId).exec().then(result => {
+			if (!result || !result.schoolId) return Promise.reject(errors.NotFound("User not found"));
 
-		return filePermissionHelper.checkPermissions(userId, path)
-			.then(() => {
-				return UserModel.findById(userId).exec().then(result => {
-					if (!result || !result.schoolId) return Promise.reject(errors.NotFound("User not found"));
+			const awsObject = createAWSObject(result.schoolId);
+			let params = {
+				Bucket: awsObject.bucket,
+				Key: path,
+				Expires: 60
+			};
 
-					const awsObject = createAWSObject(result.schoolId);
-					let params = {
-						Bucket: awsObject.bucket,
-						Key: path,
-						Expires: 60
-					};
+			if (action === 'putObject') params.ContentType = fileType;
 
-					if (action == 'putObject') params.ContentType = fileType;
-
-					return promisify(awsObject.s3.getSignedUrl, awsObject.s3)(action, params)
-						.then(res => Promise.resolve({
-							url: res,
-							header: {
-								"Content-Type": fileType,
-								"x-amz-meta-path": dirName,
-								"x-amz-meta-name": fileName,
-								"x-amz-meta-thumbnail": "https://schulcloud.org/images/login-right.png"
-							}
-						}));
-				});
-			});
+			return promisify(awsObject.s3.getSignedUrl, awsObject.s3)(action, params)
+				.then(res => Promise.resolve({
+					url: res,
+					header: {
+						"Content-Type": fileType,
+						"x-amz-meta-path": dirName,
+						"x-amz-meta-name": fileName,
+						"x-amz-meta-thumbnail": "https://schulcloud.org/images/login-right.png"
+					}
+				}));
+		});
 	}
 
 	createDirectory(userId, path) {
