+++ src/services/fileStorage/utils/leadingSlashHelper.js
@@ -1,0 +1,4 @@
+module.exports  = path => {
+	if (path[0] === '/') path = path.substring(1);
+	return path;
+};
